refactor: inline developer connection mutation options
